requirements.txt
requests>=2.28.0
tqdm>=4.64.0
python-whois>=0.7
dnspython>=2.3.0
